rename package attestations
