+++ pkg/attestation/attestation.go
@@ -1,7 +1,7 @@
 // Copyright (c) Ultraviolet
 // SPDX-License-Identifier: Apache-2.0
 
-package attestations
+package attestation
 
 import (
 	"encoding/json"
