+++ pkg/attestation/emptyprovider.go
@@ -1,7 +1,7 @@
 // Copyright (c) Ultraviolet
 // SPDX-License-Identifier: Apache-2.0
 
-package attestations
+package attestation
 
 import cocosai "github.com/ultravioletrs/cocos"
